requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
etaprogress==1.1.1
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.4
frozenlist==1.5.0
fsspec==2024.10.0
git-filter-repo==2.47.0
grpcio==1.68.1
imageio==2.36.1
imgaug==0.4.0
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.8
lazy_loader==0.4
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
piq==0.8.0
propcache==0.2.0
protobuf==5.29.1
ptflops==0.7.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.1.0
shapely==2.0.6
six==1.17.0
stonky==1.4.0
sympy==1.13.1
teletype==1.3.4
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tifffile==2024.9.20
torch==2.5.1
torchsummary==1.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
Werkzeug==3.1.3
wheel==0.44.0
yarl==1.17.1
